adding support for adding localization support BDRE-404

+++ plugin-manager/src/main/java/com/wipro/ats/bdre/pm/WarOperations.java
@@ -6,13 +6,15 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
 
 /**
  * Created by cloudera on 6/14/16.
  */
 public class WarOperations {
     private static final Logger LOGGER = Logger.getLogger(PluginManagerMain.class);
-    public void listOfFiles(File file,File parent,String md) throws IOException{
+
+    public void listOfFiles(File file,File parent,String md,String localizationFile) throws IOException{
         LOGGER.info("entering directory name : " + file.getName() + " The absolute location of dir is : " + file.getAbsolutePath());
         for(File file1 : file.listFiles()){
             if(file1.isDirectory()){
@@ -27,9 +29,9 @@ public class WarOperations {
                     new File(webappPath).mkdir();
                 }
                 if ("mdui".equals(md)) {
-                    listOfFiles(file1, parent,"mdui");
+                    listOfFiles(file1, parent,"mdui",localizationFile);
                 }else{
-                    listOfFiles(file1, parent,"mdrest");
+                    listOfFiles(file1, parent,"mdrest",localizationFile);
                 }
             }else{
                 String relativePath = file1.getAbsolutePath().replace(parent.getAbsolutePath() ,"");
@@ -40,18 +42,49 @@ public class WarOperations {
                     webappPath = System.getProperty("user.home") + "/bdre/lib/webapps/mdrest" + relativePath;
                 }
                 if (new File(webappPath).exists() && ! Files.isSymbolicLink(new File(webappPath).toPath())){
-                       continue;
-                }else {
-                    Path linkPath = new File(webappPath).toPath();
-                    if(new File(webappPath).exists()){
-                        new File(webappPath).delete();
+                    if(relativePath.contains(localizationFile)){
+                        try {
+                            Path sourcePath = new File(file1.getAbsolutePath()).toPath();
+                            byte[] data = Files.readAllBytes(sourcePath);
+                            Path targetPath = new File(webappPath).toPath();
+                            Files.write(targetPath, data, StandardOpenOption.APPEND);
+                            Path authLocalizationFile = new File(webappPath.replaceAll("/mdui", "/auth")).toPath();
+                            Files.write(authLocalizationFile, data, StandardOpenOption.APPEND);
+                        }catch (IOException io) {
+                            LOGGER.error(io + " : " + io.getMessage());
+                            throw io;
+                        }
+                    }else{
+                        continue;
                     }
-                    Path targetPath = new File(file1.getAbsolutePath()).toPath();
-                    try {
-                        Files.copy(targetPath,linkPath);
-                    } catch (IOException io) {
-                        LOGGER.error(io + " : " + io.getMessage());
-                        throw io;
+                }else {
+                    if(relativePath.contains(localizationFile)){
+                        try {
+                            Path sourcePath = new File(file1.getAbsolutePath()).toPath();
+                            byte[] data = Files.readAllBytes(sourcePath);
+                            String fileName = file1.getAbsolutePath().substring(0,file1.getAbsolutePath().lastIndexOf("."));
+                            String language = fileName.substring(fileName.lastIndexOf("_"),fileName.length());
+                            webappPath=webappPath.replaceAll("_[a-z][a-z][.]","_"+language + ".");
+                            Path targetPath = new File(webappPath).toPath();
+                            Files.write(targetPath, data, StandardOpenOption.APPEND);
+                            Path authLocalizationFile = new File(webappPath.replaceAll("/mdui", "/auth")).toPath();
+                            Files.write(authLocalizationFile, data, StandardOpenOption.APPEND);
+                        }catch (IOException io) {
+                            LOGGER.error(io + " : " + io.getMessage());
+                            throw io;
+                        }
+                    }else {
+                        Path targetPath = new File(webappPath).toPath();
+                        if (new File(webappPath).exists()) {
+                            new File(webappPath).delete();
+                        }
+                        Path sourcePath = new File(file1.getAbsolutePath()).toPath();
+                        try {
+                            Files.copy(sourcePath, targetPath);
+                        } catch (IOException io) {
+                            LOGGER.error(io + " : " + io.getMessage());
+                            throw io;
+                        }
                     }
                 }
             }
